+++ src/test/version.test.cc
@@ -2,7 +2,12 @@
 #include <gtest/gtest.h>
 
 #include <parakeet-crypto/version.h>
+#include <iostream>
 
 TEST(VersionTest, ItShouldPopulateVersion) {
-  parakeet_crypto::get_libparakeet_version();
+  auto lib_version = parakeet_crypto::get_libparakeet_version();
+  auto lib_full_version = parakeet_crypto::get_libparakeet_full_version();
+
+  std::cerr << "lib_version: " << lib_version << std::endl;
+  std::cerr << "lib_full_version: " << lib_full_version << std::endl;
 }